Experiment with malloc and aligned_alloc
Even if memory is dynamically allocated using malloc AND the data
structure definition is aligned on a specific memory boundary it
allocates a block of memory which address is a multiple of the alignment
specified in the data structure.

Removing the alignment from the data structure declaration makes
`malloc` switch back to the normal behavior, i.e. allocate a block of
memory which address is aligned on the largest data type in the
structure.

Different things happen when removing the alignment attribute from the
data structure declaration and allocating memory dynamically by using
the `aligned_alloc()` function.

In this case the memory is not allocated.