Removed `o_end` in init arr-of-orders
Before:
Passing the number of allocated items to the function that initializes
the array of orders; also, value returned by
`_order_aligned_alloc_array_of_orders` function not used.

After:
Return the pointer to the last requested item in the array when memory
is allocated; then, pass this pointer to the function that initializes
the array of orders. The latter function no longer has to compute the
pointer to the last requested item.