2x speed optimization
Loops have been rearranged.

Instead of testing every iterations if the induction variable is LESS
THAN, which implies:

1 - load the variable into a register
2 - subtract the max number of iterations
3 - branch?

The conditions have been replaced with a test on the variable itself
that only implies a load.

Also, the test condition has been moved at the end of the loop.

Do-while are used instead of for or while loops.

There are no if statement. This helps to avoid stalls in the pipeline
and creating bubbles, since the IF STATEMENT implies a jump to the block
of the if statement AND also evaluating a condition.

The conditions are instead evaluated directly in the while at the end of
the loop.