Increasing s = 8 on the column: worst performance
This change was made to minimize the number of moves inside a cache
line:

Each item is 4 bytes;
The cache line size is 128 bytes, 32-times the size of a single item;
By increasing the shift to 8 and using 4 variables at the same time to
span the columns in a row of the matrix the items covered in a single
cycle are exactly 32.

This decreases the number of cycles x cache line.

It means that a cache line is loaded in the cache memory and after 8
cycles is evicted.

Using shift = 4 it tooks 2-times 4-cycles (a total of 8 cycles) to evict
the same cache line.

It may be possible that this does not allow the PREFETCH UNIT to
successfully prefetch the next cache line because there is A LESS
EVIDENT PATTERN, since the iterations are 8 straightforward instead of 2
cycles of 4 inner iterations.