Rearrange col counter declarations based on access
This seems not to improve performance.

When loading the addresses to be assigned there are two main cache lines
that are hit:

[sp, #48] and [sp, #40]

These are the memory address of the current row and the memory address
of the shifted row.

Before the code prefers the same cache line to the same column counter;
Now the code regroup the assignments based on the same column counter.

This is not better, it seems slightly worst.