Do statement generates a better assembly!
In this version I add the same instructions packed inside a do statement
instead of the for statement.

The compiler is capable of generating assembly instructions for
incrementing both `sum` and `i` in the same basic block.

This is the same effect that Mike Acton shows in his article about
"Branch elimination" at:
https://web.archive.org/web/20071224201329/http://www.cellperformance.com/articles/2006/04/programming_with_branches_patt.html

In the for loop example the compiler needs to place the increment of the
sum variable in an intermediate basic block.

The limitation on performance in the for loop case is that the CPU may
be less efficient in executing instructions out-of-order.

In the example with the do statement since the load, add and store
instructions for both sum and i variables reside in the same basic
block, the CPU may be more effective at execution.