Infinite loop + if: valid alternative to do-while
In this version I add an implementation of the same loop that is very
similar to the do statement: an infinite loop (i.e., not conditional
branch when entering the loop) with an if statement inside that breaks
the loop in case the exit condition is met.

The program is compiled WITHOUT optimization flags and the assembly
instructions generated contain useless branch to empty blocks.

This gives even more room for optimization!