The compiler tricked me!!
Compile the code from the previos commit with the O2 optimization flag
and the assembly instructions generated for the functions are EMPTY.

There is nothing inside the functions.

Why? Because no other part of the code uses the result computed by those
functions, the compiler just generates empty functions.

Interesting behavior to better understand how compiler generates code:

THEY CONSIDER THE EFFICIENCY IMPACT.

Being efficient means MINIMIZE the waste of resources.

In this specific environment the goal is to minimize power consumption!!