Compiler tricks me again!!
Since in the previous version the compiler generates functions with only
a return statement in their bodies, I decide to return sum back to the
caller to push the compiler to generate the code.

However, the compiler is so clever that optimizes away the loops,
precomputes the sum based on the total number of iteration since it is
static and directly returns it.