The advantage of inifinite + if break loops!
When compiling the code with optimization flags off, it is clear the
effect the different loop implementations have on the respective
assembly instructions generated.

=============
The for loop:
=============

The major limitation of the for loop that is possible to spot from the
assembly instructions is due to the fact that the loop is split into
three different blocks of code + a pre-loop header:

The pre-header is used to initialize the variables and is common to all
the other loop implementations.

The first block of code contains the assembly instructions for testing
the condition. Note that this is the first block of the loop due to the
nature of the for loop itself.

The second block is the body of the loop.

The third block is the update of the induction variable.

By looking at the code generated by the compiler the three blocks are
CRYSTAL CLEAR.

The problem is that having many blocks implies many transitions from one
block to another.

This works against the modern CPUs architectures in particular for
out-of-order execution and register renaming.

=================
The do-while loop
=================

There is an advantage in using a do statement for repeating the same
instructions over and over again: the update of the induction variable
is in the body of the loop.

This gives the CPU more power on reording the instructions and using
register renaming technique to improve the performance since now the
assembly instructions generated are divided into two basic blocks + a
pre-header:

Again, the pre-header is common to any of the loop and serves to perform
various initializations.

The first block is the body of the loop.

The second is the block that contains the instructions for the
conditional branch execution and jumps back to the top of the first
block if the jump is taken.

This can be easily achieved even with a for loop, but the resulting loop
should no longer be considered a true for loop! It's more similar to a
while.

============================
The infinite + if-break loop
============================

This is even more processor-friendly since it moves the exit condition
DIRECTLY IN THE BODY OF THE LOOP, hence making all the assembly
instructions related to the loop in one basic block.

This is even better because the CPU may take advantage of the fact that
the code is linear and there are no jumps between basic blocks of code
and it may prefetch the instructions, move on the pipeline, execute
out-of-order, use register renaming, etc.

This empowers a modern CPU.

For loops are deprecated on modern hardware.

(In reality they are not, simply because by using the optimization flags
the compiler is capable of generating the same assembly instructions
starting from the three different implementations, but let's assume for
a minute the compiler is dumb and cannot do optimization... in that case
the for loop will be bad).