Benchmark test1: basic nested for over rows.
In this version I add the most straightforward function to just
initialize a 2D array by doing some arithmetic computations.

What I noticed by looking at the assembly instructions generated by the
compiler is that by using the `for` loop the compiler splits the code
into 3 basic blocks.

Having more basic blocks means the CPU need to jump to those blocks
during the execution.

Each basic block can be thought of like an island: the CPU has to take a
flight (i.e., jump) to move to the basic block and start executing the
instructions inside.

This is not as efficient as it could be, especially on modern hardware
because the scheduler has a more difficult job for scheduling the
instructions.

As a consequence, out-of-order execution CPU's capabilities cannot be
fully exploited.