Replace fdiv with fmul by the inverse is faster!
In this version I extract the floating-point division by 3.0f since it
is a constant and I pre-compute the inverse of 3.0 and stored in a
variable `c` outside the outermost loop to avoid recomputing every time
the loop is entered the same quantity.

The result is a significant improvement in the performance.