Take advantage of associative prop of mult!
In this version I split the arithmetic expression:

	(i+j) * c

Into three different computations:

	i * c
	j * c

And the third arithmetic operations is the sum of the preceding two.

This means that two out of three operations are independent and can be
done in parallel by the processor.

The third operations can be prefetched by the processor and as soon as
the two precidings operations are done it can be executed.

This makes use of more memory but by explicitly EXTRACT the arithmetic
operations allows the compiler to produce optimized code.