Reverse walk the array produces more instructions!
In this version I test the effect of walking backward the elements in a
row by initializing j to ncols - 1 and just test the condition on the
value of j without having to compute at every iteration the difference
between the new value of j and ncols.

Moreover, the test condition needs to wait for the result of j = j + 1
before computing the difference and this may slow down the CPU pipeline.

That's the rationale behind this test.

However, by looking at the assembly code the effect of the reverse walk
was to generate two more instructions for every 128-bit vector register
operation.