Outer loop turned into inf + if break
In this version I turned the outer loop into an infinite loop with an if
statement at the bottom to test the condition for exiting the loop.

The assembly generated is not significantly different from the other
versions meaning that the compiler already has done this kind of
`optimization`.

Another important thing: I move outside the inner loop the computation
of `i * c` since the quantity only depends on `i` and not on the inner
loop induction variable `j` BUT WHAT I FOUND by looking at the assembly
generated instructions is that the compiler already had done this
optimization!!