Do not declare local variable for casting values!
I removed the `_tmp` variable declared inside the outermost loop where I
stored the cast value of `_ii` and directly inline the expression in the
computation of `_res` but the effect on the generated assembly
instructions is nothing: the compiler already strip out this
intermediate value because it consider it useless.

This is a good example to understand how far to push yourself in trying
to write simple and small traces to help the compiler and here the
compiler said: "Hey, it's okay, I got it, don't need this!".