Bitwise op unfold inner loop for compact code!
In this version I test a new way of walking throught the 2D array by
using a single loop and bitwise mask to update the row and column
indices.

I pre-compute the upper limit (considering the overflow risks) and test
the condition on the index variable.

The code is significant more compact than all the previous versions but
more than 2x slower!!

Need loop unrolling to make parallel computations inside the loop and
speed-up the program!!