Mul instruction is not the limitation!
Try to hardcode the product between `nrows` and `ncols` in the
`_loop_test7()` function but this does not significantly improve the
result.

This confirm the hypothesis that the main limitation is not given by the
initial multiplication but by the fact that the loop IS NOT VECTORIZED,
hence we are not taking advantage of the CPU power of executing multiple
instructions in parallel!