Makefile DOES NOT WORK!!
In the previous version I tried to generate the executable file by using
`clang` for generating assembly files, use them to generate object files
needed to generate the final executable, but this does not work.

Including a simple `echo` in the list of command for each of the target
makes me understand that running `clang` run all the stages at once and
does not generate the object files from the assembly files.

To correctly do that you need to enter the manual mode and use `as` and
`ld`.

The problem is that the linker gives an error that says:

"ld: library not found for -lSystem"

since it links to dynamic libraries stored in cache to speed-up the
process and is not the correct way of doing it.

This version serves just as a reminder to my future self to understand
the kind of errors and the procedure to manually generate an executable
file.