Extract buyer module
The buyer module defines its interface in the buyer.h file and the
implementation in the buyer.c file.

This module contains the functions related to the basic buyers
information.

The order module defines the logic and interface to manage a specific
information about buyers: orders.

Currently the order module supports paid and unpaid orders only.

In future it may be extended to include other kinds of orders.

The order module should be seen as a plugin to the buyer module.

If there is the need to manipulate other information about a single
buyer, the buyer module can be easily extended by defining another
module specific to the information needed at the time.

This makes the program more manageable and clearly defines the
interfaces and responsibilities of each subportion of the project.