Tweak optim for an even fast vector2_add!
First I tried to change the if(...){continue;} to the an if(value){...}
and this makes the assembly generated by the compiler much better.

The next thing was to extract the sum of v1 and v2 items into local
variables and this allows for even more compact code generated by the
compiler BUT MORE IMPORTANTLY the instructions are arranged better:

LOAD first, then update/compute, then store.