Sum rows optim 1: extract `sum` local var
In this version I extract a local variable `sum` initialized to zero to
store the sum of the items over the columns in the vector2 object.

This reduces the number of assembly instructions generated by the
compiler.

Also, this is more parallel-friendly since each thread has its own copy
of the stack, using a local variables frees the threads to do work
independently instead of continuing to update each other on the values
of the variable.