Different inner loop impl in vector2_add func
I try to use 4 local induction variable instead of 4 local pointers to
rows to test the differences in the assembly instructions generated by
the compiler.

Compared to the version that uses pointers the compiler generates more
instructions, especially more load instructions and the code seems
slightly slower with 1000x1000 benchmark.

Need to test on larger data structures, but I think it's preferable the
version that uses pointers.

In the latter case, it's possible to use the "restrict" keyword on
pointers to allow the compiler to make further optimizations.

The same semantic cannot be achieved by using 4 induction variables.