vector2 mul j-loop optimization
In this version I change the innermost loop in my version of the vector2
multiplication function by using an infinite loop combined to an if
statement to test the exit condition.

The induction variable j has been replaced by the pointers: instead to
increment the induction variable the pointer is incremented.

There is a new variable to control the exit condition from the loop.

The variable is completely independent from anything else, allowing the
processor to schedule instructions as better as it can.

The variable is initialized to the maximum number of iterations and is
decremented until 0.

This simplified the test condition because the processor does not have
to compute at every single iteration the difference between the updated
value of the induction variable and the max number of iterations.