refactor mid-loop in vector2 mul function
In this version I apply the same changes applied to the j loop in the
vector2 mul function to the loop over k induction variable.

There is no longer an induction variable k, but there is a kiter
variable initialized at the maximum number of iterations and is
decremented at every time.

The loop is an infinite while with an if statement to control the exit
condition.

The kiter variable is completely independent from other variables in the
loop and this allows the processor to better schedule instructions.

The induction variable k has been removed in favor of directly increment
the pointers!

Moving to the next line == add the number of columns.
Moving to the adjacent item on the same row == add 1.