Introducing jiter induction variable
In this version I add the `jiter` induction variable.

This is completely decoupled from the `j` induction variable.

The latter is used to index the vector2_t objects.
The former is used to test the termination condition.

By inspecting the assembly instructions generated by the compiler the
instructions in the innermost loop are layed out by following the
scheme:

LOAD -> COMPUTE -> STORE