Removing local var `j` makes code ~2x slower!!!
I try to remove the local variable `j` in favor of using the pointers
and directly incrementing them instead of having an induction variable
used to index the arrays.

Without j the code was ~2x slower!!

This is because modern processors have many cores and having local
variables is an advantage because it allows the threads running on cores
to be independent.

Also, local variable j is stored on the stack and each thread has its
own copy of the stack, this means decoupling the execution on multiple
cores and avoid the needs to update cache lines that may be invalidate
by other cores.