EXP1: refactor 3-level nested loops in vector2 mul
This version introduces changes in the implementation of the 3-level
nested loop in vector2 mul function.

The idea behind this approach is to reduce the latency by precomputing
the values EXPLICITLY before storing them.

This version heavily relies on the LOAD - UPDATE - STORE pattern.

Another significant change is the moving of the update at the tail of
the loop BUT BEFORE the testing condition.

The intention behind this approach is to allow the compiler to generate
assembly instructions without branches.

In this way the values are always updated before the testing condition,
hence this empowers the compiler to generate more compact assembly
instructions.